+++ foxes/input/yaml/windio/read_attributes.py
@@ -167,7 +167,13 @@ def _read_blockage(blockage_model, induction, algo_dict, mbook, verbosity):
         print("      Name:", wname)
         print("      Contents:", [k for k in blockage_model.keys()])
     if wname not in ["None", "none"]:
-        indc_dict = Dict(wmodel_type=indc_def_map[wname], induction=induction)
+        kys = list(blockage_model.keys())
+        for k in kys:
+            if len(k) > 3 and k[:3] == "ss_":
+                blockage_model[k[3:]] = blockage_model.pop_item(k)
+        indc_dict = Dict(
+            wmodel_type=indc_def_map[wname], induction=induction, **blockage_model
+        )
         mbook.wake_models[wname] = WakeModel.new(**indc_dict)
         if verbosity > 2:
             print(f"      Created wake model '{wname}':")
@@ -378,7 +384,7 @@ def read_attributes(wio_attrs, idict, mbook, verbosity=1):
             print("    Reading flow_model")
             print("      Name:", fmname)
             print("      Contents:", [k for k in flow_model.keys()])
-        if fmname != "foxes":
+        if verbosity > 0 and fmname != "foxes":
             print(f"Running flow model 'foxes', overruling original choice '{fmname}'")
 
     # read analysis: