+++ foxes/models/farm_controllers/__init__.py
@@ -3,3 +3,4 @@ Farm controller models.
 """
 
 from .basic import BasicFarmController as BasicFarmController
+from .op_flag import OpFlagController as OpFlagController